refactor: move layout function to separate file
